[FIX] adding symlink checks to avoid unexpected behaviour
Any file that is on Template path instead of tmp_dir path now gets checked to see if the said file is a symlink or an actual file.
This is not done for Files or some other parameters as they cannot be stored inside the Template and can only be in tmp_dir which is assumed to use non-symlink paths.
closes #9